loadbalancer: modified avg_time calculation, int->long
git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@2993 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ module-stat.c
@@ -125,10 +125,10 @@ void calc_stat(READER_STAT *stat)
 {
 	int i;
 	int c=0;
-	int t = 0;
+	long t = 0;
 	for (i = 0; i < MAX_STAT_TIME; i++) {
 		if (stat->time_stat[i] > 0) {
-			t += stat->time_stat[i];
+			t += (long)stat->time_stat[i];
 			c++;
 		}
 	}
